Add Service operations interface, rename and move implementation of service operations and move to other package
